refactor: rearrange errors positioning for cleaner syntax
